% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attrition_drop.R
\name{attrition_drop}
\alias{attrition_drop}
\title{Replace imputations with NAs if respondent did not participate in a survey wave}
\usage{
attrition_drop(imp_df, wave, wave_indicator)
}
\arguments{
\item{imp_df}{Data frame from after imputation}

\item{wave}{The wave number}

\item{wave_indicator}{A logical vector. Each element should be
TRUE if the respondent is in the wave and FALSE if not.}
}
\value{
Updated \code{imp_df}
}
\description{
Replace imputations with NAs if respondent did not participate in a survey wave
}

