% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_matrix.R
\name{bounds_matrix}
\alias{bounds_matrix}
\title{Make bounds matrix for amelia}
\usage{
bounds_matrix(varlist)
}
\arguments{
\item{varlist}{list returned by \code{\link{imp_varlist}}}
}
\value{
A matrix that can be passed to the \code{bounds} argument of
  \code{amelia()}.
}
\description{
Make bounds matrix for amelia
}

