% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap.R
\name{cap}
\alias{cap}
\title{Cap}
\usage{
cap(x, at)
}
\arguments{
\item{x}{A numeric vector or matrix.}

\item{at}{Value at which to cap \code{x}.}
}
\value{
\code{x} but with values greater than \code{at} set equal to \code{at}
}
\description{
Cap
}
\examples{
cap(1:5, at = 2)
}

