% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_numeric_leftclosed.R
\name{cut_numeric_leftclosed}
\alias{cut_numeric_leftclosed}
\title{Wrapper for \code{cut()} with \code{labels = FALSE}, \code{right = FALSE}}
\usage{
cut_numeric_leftclosed(x, breaks)
}
\arguments{
\item{x, breaks}{same as \code{\link[base]{cut}}.}
}
\value{
An integer vector of level codes. As with \code{cut}, values which
  fall outside the range of breaks are coded as NA (as are NaN and NA
  values).
}
\description{
Integer codes are returned instead of a factor. Intervals closed on left and
open on right.
}
\examples{
x <- 1:10
y <- cut_numeric_leftclosed(x, breaks = c(0, 2, 7))
cbind(x, y)

z <- cut_numeric_leftclosed(x, breaks = c(1, 2, 7, Inf))
cbind(x, z)

}
\seealso{
\code{\link[base]{cut}}
}
