% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac2binary.R
\name{fac2binary}
\alias{fac2binary}
\title{Convert 2-level factor variable to a binary 0/1 variable}
\usage{
fac2binary(x)
}
\arguments{
\item{x}{A 2-level factor variable}
}
\value{
\code{x}, converted to a binary variable
}
\description{
Convert 2-level factor variable to a binary 0/1 variable
}

