% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_names.R
\name{flag_names}
\alias{flag_names}
\title{Get names of imputation flag variables}
\usage{
flag_names(vars, flag = "pimp.", waves = 1:12, year = NA)
}
\arguments{
\item{vars}{Character vector of variable names}

\item{flag}{The text to use when creating the names of flag variables. If
not baseline, then "qx" will be prepended to \code{flag}, where "x" is
the appropriate wave.}

\item{waves}{A numeric vector indicating any survey waves beyond baseline for
which variable names should be looked for in \code{vars} and corresponding
flag names returned}
}
\value{
Character vector of names for the imputation flag variables
  corresponding to vars
}
\description{
Get names of imputation flag variables
}
\examples{
vars <- c("bob", "q4bob", "q8bob", "jim", "q4jim")
flag_names(vars)
flag_names(vars, waves = 4) # q8 vars treated like baseline because 8 not in waves

}
