% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imputation_functions.R
\name{imputation}
\alias{Mode}
\alias{gateway_zeros}
\alias{imp_reduce}
\alias{imputation}
\alias{nm}
\alias{par_mice}
\title{Functions used for multiple imputation}
\usage{
par_mice(df, impmthd, prd_mat)

imp_reduce(all_imps)

gateway_zeros(x, y, df)

Mode(x)

nm(x)
}
\arguments{
\item{df}{a \code{data.frame}.}

\item{impmthd}{a vector of strings with length ncol(data), specifying the
elementary imputation method to be used for each column in data imputation
methods. See \code{\link[mice]{mice}}.}

\item{prd_mat}{a square binary matrix. See \code{\link[mice]{mice}}.}

\item{all_imps}{data for all imputations}
}
\description{
Functions used for multiple imputation
}

