% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_continuous_from_category.R
\name{impute_continuous_from_category}
\alias{impute_continuous_from_category}
\title{Impute continuous value given a bracket (i.e., bounds)}
\usage{
impute_continuous_from_category(contvar, catvar, df, breaks)
}
\arguments{
\item{contvar}{Name of the continuous variable with missing values}

\item{catvar}{Name of the categorical variable indicating the bracket}

\item{df}{The data frame in which to look for \code{contvar} and \code{catvar}}

\item{breaks}{The breakpoints for all the brackets}
}
\value{
An updated version of \code{df} including the imputations
}
\description{
Used for naive imputation of earnings (and income from other family members
in HH) for respondents who provided a bracket but not the actual value. This
draws a value from a uniform distribution with the bounds implied by the
bracket.
}

