% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_waves.R
\name{include_waves}
\alias{include_waves}
\title{Take baseline variable names and return variable names for baseline and/or
later waves}
\usage{
include_waves(vars, waves = 0)
}
\arguments{
\item{vars}{A character vector of (baseline) variable names}

\item{waves}{A numeric vector of survey wave numbers (0 for baseline)}
}
\value{
\code{vars} updated to include the the names of the corresponding
  variables in the waves indicated by \code{waves}.
}
\description{
Take baseline variable names and return variable names for baseline and/or
later waves
}
