% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator12_to_indicator01.R
\name{indicator12_to_indicator01}
\alias{indicator12_to_indicator01}
\title{Convert (one or more) 1/2 indicator variables to 1/0 indicator variables}
\usage{
indicator12_to_indicator01(vars, df)
}
\arguments{
\item{vars}{A character vector of variable names}

\item{df}{The data frame in which to look for \code{vars}}
}
\value{
\code{df}, with \code{vars} converted to 0/1 indicators (binary
  variables)
}
\description{
NOTE: This converts the 2s to 0s and leaves the 1s and 1s!!
}
