% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margins_for_rake.R
\name{margins_for_rake}
\alias{margins_for_rake}
\title{Sample and population margins for raking}
\usage{
margins_for_rake(me_vars, int_vars, interact_with, design)
}
\arguments{
\item{me_vars}{character vector of variable names for main effects.}

\item{int_vars}{character vector naming variables to be interacted with
\code{interact_with}.}

\item{interact_with}{character string naming the variable with which
the variables in \code{int_vars} will be interacted. Currently
\code{interact_with} can only be a single variable.}

\item{design}{survey object. See \code{\link[survey]{svydesign}}.}
}
\value{
A list with components \code{samp} and \code{pop}. \code{samp} is a
  list of formulas describing sample margins. \code{pop} is a list of
  data.frames and tables describing the corresponding population margins. See
  the description of the \code{sample.margins} and \code{population.margins}
  arguments for \code{\link[survey]{rake}}.
}
\description{
This function can be used to construct the \code{sample.margins} and
\code{population.margins} arguments for \code{\link[survey]{rake}}.
}
