% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/max_over_min.R
\name{max_over_min}
\alias{max_over_min}
\title{Ratio of maximum to minimum}
\usage{
max_over_min(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
\code{max(x) / min(x)}
}
\description{
Ratio of maximum to minimum
}

