% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mean_plus_sds.R
\name{mean_plus_sds}
\alias{mean_plus_sds}
\title{Mean plus \eqn{k} standard deviations}
\usage{
mean_plus_sds(x, k)
}
\arguments{
\item{x}{numeric vector.}

\item{k}{number of standard deviations.}
}
\value{
\code{mean(x) + k * sd(x)}
}
\description{
Mean plus \eqn{k} standard deviations
}

