% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_continuous_from_category.R
\name{naive_imputation_step}
\alias{naive_imputation_step}
\title{Naively impute continuous value given a bracket (i.e., bounds)}
\usage{
naive_imputation_step(df, contvars, catvars, breaks)
}
\arguments{
\item{df}{The data frame to use.}

\item{contvars}{Character vector of names of the continuous variables with
missing values.}

\item{catvars}{Character vector of names of categorical variables indicating
the bracket.}

\item{breaks}{The breakpoints for all the brackets.}
}
\value{
An updated \code{df} including the imputations.
}
\description{
Naive imputation of earnings (and income from other family
  members in HH) for respondents who provided a bracket but not the actual
  value. This draws a value from a uniform distribution with the bounds
  implied by the bracket.
}
