% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAs_to_zeros.R
\name{NAs_to_zeros}
\alias{NAs_to_zeros}
\title{Convert NAs in imputation flags to zeros}
\usage{
NAs_to_zeros(df, patt = "pimp.")
}
\arguments{
\item{df}{A data frame}

\item{patt}{Pattern to \code{grep} for. Used to find the column names
corresponding to the imputation flag variables.}
}
\value{
\code{df}, updated
}
\description{
This function takes a data.frame \code{df} and converts all \code{NA}s in
the imputation flag variables to zeros.
}

