% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize to have mean 1}
\usage{
normalize(x)
}
\arguments{
\item{x}{vector to normalize to have mean 1.}
}
\value{
\code{x / mean(x)}
}
\description{
Normalize to have mean 1
}
