% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ones.R
\name{ones}
\alias{ones}
\title{Make a one-column matrix of 1s}
\usage{
ones(sample_size)
}
\arguments{
\item{sample_size}{The number of rows for the matrix.}
}
\value{
\code{matrix(1, sample_size, 1)}
}
\description{
Make a one-column matrix of 1s
}

