% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_imp_NAs.R
\name{prepare_imp_NAs}
\alias{prepare_imp_NAs}
\title{Make all values of vars with imputation flag = 1 into NAs}
\usage{
prepare_imp_NAs(df, vars, flag = "pimp.", waves = 1:8)
}
\arguments{
\item{df}{The data.frame to use}

\item{vars}{Character vector of variable names}

\item{flag, waves}{See \code{\link{flag_names}}}
}
\value{
A modified version of df with all flagged values of vars made NA
}
\description{
Make all values of vars with imputation flag = 1 into NAs
}

