% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_weights_to_pop.R
\name{rescale_weights_to_pop}
\alias{rescale_weights_to_pop}
\title{Rescale weights to sum to population}
\usage{
rescale_weights_to_pop(x, pop_size)
}
\arguments{
\item{x}{vector of weights}

\item{pop_size}{population size}
}
\value{
a vector that sums to \code{pop_size}
}
\description{
Rescale weights to sum to population
}

