% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_weights.R
\name{save_weights}
\alias{save_weights}
\title{Save weights as .rds or .dta file}
\usage{
save_weights(data, format = c("rds", "dta"), file_name = paste0("weights.",
  format), include_reps = TRUE)
}
\arguments{
\item{data}{the \code{data.frame} to take the weights from.}

\item{format}{save as .rds file (R) or .dta file (Stata).}

\item{file_name}{the path to where the file should be saved.}

\item{include_reps}{logical. Should replicate weights by saved if found?}
}
\value{
a .dta or .rds file, depending on \code{format}.
}
\description{
This will save all variables with names containing 'weight_'
}
