% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_impute.R
\name{simple_impute}
\alias{simple_impute}
\title{Simple imputations}
\usage{
simple_impute(n, vals, weight_var)
}
\arguments{
\item{n}{The number of missing values to impute.}

\item{vals}{The values to sample from (with replacement).}

\item{weight_var}{Variable from which to compute relative frequencies of \code{vals}
to use as \code{prob} argument to \code{\link[base]{sample}}. If missing then
the values in \code{vals} are sampled with equal probability.}
}
\value{
A vector of length \code{n}.
}
\description{
Wrapper for \code{\link[base]{sample}}
}

