% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/survey_summary.R
\name{survey_summary}
\alias{survey_summary}
\title{Print survey summary statistics}
\usage{
survey_summary(x, which = "mean", design, msg = NULL)
}
\arguments{
\item{x}{variable to use.}

\item{which}{character string. The function paste0("svy", which) will be used.}

\item{design}{survey design object.}

\item{msg}{character string. An optional message to print.}
}
\description{
Print survey summary statistics
}

