% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/upper_quantile_trimming.R
\name{upper_quantile_trimming}
\alias{upper_quantile_trimming}
\title{Trim high survey weights using quantile as upper bound}
\usage{
upper_quantile_trimming(design, prob, ...)
}
\arguments{
\item{design}{A survey design object.}

\item{prob}{A probability in \eqn{[0,1]}}

\item{...}{Additional arguments to pass to \code{\link[survey]{trimWeights}}}
}
\value{
See \code{\link[survey]{trimWeights}}
}
\description{
Wrapper for \code{\link[survey]{trimWeights}}
}

