% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wave_weights.R
\name{wave_weights}
\alias{wave_weights}
\title{Compute weights for (post-baseline) RH survey waves}
\usage{
wave_weights(data, wave, sample.margins, population.margins,
  replicates = FALSE, strict_trimming = TRUE, acs_child_data = NULL)
}
\arguments{
\item{data}{\code{data.frame} to use (the RH data)}

\item{wave}{numeric. The wave for which to compute weights.}

\item{sample.margins}{a list.}

\item{population.margins}{a list.}

\item{replicates}{logical. Should replicate weights also be computed?}

\item{strict_trimming}{logical.}

\item{acs_child_data}{ACS data for children (age < 18). Only used if
\code{wave = 1} and child weights should to be computed. If
\code{acs_child_data} is not \code{NULL} and \code{wave == 1} child weights
will be computed. If \code{acs_child_data} is \code{NULL} or \code{wave !=
1} then no child weights will be computed.}
}
\value{
The \code{data.frame} provided for the \code{data} argument will be
  returned will additional columns for the computed weights.
}
\description{
Compute weights for (post-baseline) RH survey waves
}

