% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weights_hist.R
\name{weights_hist}
\alias{weights_hist}
\title{Histogram of weights}
\usage{
weights_hist(w, title = NULL)
}
\arguments{
\item{w}{vector of weights.}

\item{title}{character string. The title for the plot.}
}
\description{
Histogram of weights
}

