% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amelia2df.R
\name{amelia2df}
\alias{amelia2df}
\title{Retrieve an imputed dataset (as a data frame) from an amelia object}
\usage{
amelia2df(x, which_imp = 1)
}
\arguments{
\item{x}{An object of class \code{"amelia"} obtained from running
the \code{\link[Amelia]{amelia}} function.}

\item{which_imp}{If \code{x} contains multiple imputed datasets (i.e., if
\code{m} argument to \code{amelia} was greater than 1), which of the
imputed datasets should be returned?}
}
\value{
A data.frame
}
\description{
Retrieve an imputed dataset (as a data frame) from an amelia object
}
