% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cut_numeric_leftclosed.R
\name{cut_numeric_leftclosed}
\alias{cut_numeric_leftclosed}
\title{Wrapper for \code{cut()} with \code{labels = FALSE}, \code{right = FALSE}}
\usage{
cut_numeric_leftclosed(x, breaks)
}
\arguments{
\item{x,breaks}{same as \code{\link[base]{cut}}.}
}
\value{
an integer vector of level codes.
}
\description{
Integer codes are returned instead of a factor. Intervals closed on left and
open on right.
}
\seealso{
\code{\link[base]{cut}}
}

