% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_df.R
\name{empty_df}
\alias{empty_df}
\title{Create a placeholder NA-filled data.frame}
\usage{
empty_df(nrow, ncol)
}
\arguments{
\item{nrow, ncol}{the number of rows and columns for the \code{data.frame}}
}
\value{
an \code{nrow} by \code{ncol} \code{data.frame} of NAs.
}
\description{
Create a placeholder NA-filled data.frame
}
\examples{
d <- empty_df(3, 2)
print(d)

}
