% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp_varlist.R
\name{imp_varlist}
\alias{imp_varlist}
\title{Create imputation information list}
\usage{
imp_varlist(vars, types, bounds = list(), waves = NULL)
}
\arguments{
\item{vars}{Character vector of variable names}

\item{types}{Character vector of variable types}

\item{bounds}{List of bounds, or NAs. E.g. list(NA, NA, NA, c(0, Inf))}

\item{waves}{A numeric vector of survey wave numbers indicating waves to
include in addition to baseline}
}
\value{
A list containing info on the type of variable and any bounds
}
\description{
Create imputation information list
}

