% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_waves.R
\name{include_waves}
\alias{include_waves}
\title{Add to vector of baseline variable names the names of corresponding variables
in later survey waves}
\usage{
include_waves(vars, waves = NULL)
}
\arguments{
\item{vars}{A character vector of (baseline) variable names}

\item{waves}{A numeric vector of survey wave numbers}
}
\value{
\code{vars} updated to include the the names of the corresponding
  variables in the waves indicated by \code{waves}.
}
\description{
Add to vector of baseline variable names the names of corresponding variables
in later survey waves
}

