% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ones.R
\name{ones}
\alias{ones}
\title{Make a one-column matrix of 1s}
\usage{
ones(sample_size)
}
\arguments{
\item{sample_size}{The number of rows for the matrix.}
}
\value{
A one column matrix of 1s with \code{sample_size} rows.
}
\description{
Make a one-column matrix of 1s
}
